Basket/17-12-2025-habit-tracker/habit_tracker/requirements.txt
python-telegram-bot~=13.15
python-dotenv
aiofiles
aiocron